+++ YesplanVIP/Models/Singles/Resourcebooking/InstantiableResourceUseGroup.swift
@@ -12,11 +12,11 @@ struct InstantiableResourceUseGroup: Codable {
     var url: String
     var id: String
     var _type: InstantiableResourceUseGroupType
-    var number: Int
+    var number: Double
     var resource: ResourceUseResource? // Resource Location_Id or Resource_Id or Placeholder_Id
     var children: [InstantiableResourceUse] // (instantiable resource use)*
     var event: Base? // tom // NOT Event_Id, Event_Id or Schedule
-    var actualnumber: Int?
+    var actualnumber: Double?
 }
 
 enum InstantiableResourceUseGroupType: String, Codable {
