+++ YesplanVIP/Models/Singles/Resourcebooking/ResourceSetUse.swift
@@ -16,12 +16,12 @@ struct ResourceSetUse: Codable {
     var owningteam: Usergroup?
     var owninggroup: Usergroup?
     var _type: ResourceSetUseType
-    var number: Int
+    var number: Double
     var resource: ResourceUseResource? // Resource Location_Id or Resource_Id or Placeholder_Id
     var children: [ResourcebookingChildren]
     var event: Base? // tom // NOT Event_Id, Event_Id or Schedule
     var resourceinstance: Base? // NOT Resourceinstance_Id
-    var actualnumber: Int?
+    var actualnumber: Double?
 }
 
 enum ResourceSetUseType: String, Codable {
